django_hw_4/requirements.txt
psycopg2-binary
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
